pyproject.toml
[tool.poetry]
name = "ask-intercom"
version = "0.1.0"
description = "AI-powered insights from Intercom conversations"
authors = ["Ask-Intercom Team <team@ask-intercom.ai>"]
readme = "README.md"
packages = [{include = "src"}]

[tool.poetry.dependencies]
python = "^3.11"
openai = "^1.0.0"
httpx = "^0.27.0"
python-dotenv = "^1.0.0"
pydantic = "^2.0.0"
rich = "^13.0.0"
fastapi = "^0.115.13"
uvicorn = "^0.34.3"
mcp = "^1.1.2"
httpx-sse = "^0.4.0"
aiohttp = "^3.11.11"
websockets = "^15.0.1"
# Removed: fast-intercom-mcp (now external Docker service)
# fast-intercom-mcp = {git = "https://github.com/evolsb/FastIntercomMCP.git"}

[tool.poetry.group.dev.dependencies]
pytest = "^8.0.0"
pytest-asyncio = "^0.21.0"
black = "^23.0.0"
ruff = "^0.1.0"
pytest-httpx = "^0.30.0"
pytest-mock = "^3.14.1"
requests = "^2.32.4"

[tool.poetry.scripts]
ask-intercom = "src.cli:main"
ask-intercom-interactive = "src.cli:interactive_main"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.black]
line-length = 88
target-version = ['py311']

[tool.ruff]
target-version = "py311"
line-length = 88
select = [
    "E",  # pycodestyle errors
    "W",  # pycodestyle warnings
    "F",  # pyflakes
    "I",  # isort
    "B",  # flake8-bugbear
    "C4", # flake8-comprehensions
]
ignore = [
    "E501",  # line too long, handled by black
]
